backend/requirements.txt
# ============================================================================
# ATS Resume Checker — Python Dependencies
# ============================================================================

# ── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── AI / LLM ───────────────────────────────────────────────────────────────
google-generativeai==0.8.2

# ── Resume Parsing ──────────────────────────────────────────────────────────
PyMuPDF==1.24.10
python-docx==1.1.2

# ── NLP & Text Processing ──────────────────────────────────────────────────
scikit-learn==1.5.2

# ── LaTeX Templating ───────────────────────────────────────────────────────
Jinja2==3.1.4

# ── Configuration ──────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── CORS & Utilities ───────────────────────────────────────────────────────
aiofiles==24.1.0
